% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato.R
\name{import_lato}
\alias{import_lato}
\title{Import Lato font}
\usage{
import_lato()
}
\value{
Path to folder containing Lato TTF file (invisibly)
}
\description{
First registers fonts so that they can be used with the PDF, PostScript, or
Windows bitmap output devices, then imports Lato.
}
\details{
This does not replace the need to install Lato on your system for output to
other graphics devices (e.g. png). You can do this by downloading the font
from the web and saving it to the appropriate location for your system
(\verb{/Library/Fonts/} on macOS). \code{import_lato()} invisibly returns the path to
the TTF file that's included in the package, in case you prefer to copy it
from there to your system font library.
}
\examples{
\dontrun{
import_lato()
}
}
\references{
\url{http://www.latofonts.com/}
}
