% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_image}
\alias{logo_image}
\title{Logo image}
\usage{
logo_image(
  image = system.file("SageBionetworks-Logo.png", package = "sagethemes"),
  x = 0.95,
  y = 1,
  hjust = 1,
  vjust = 1,
  halign = 1,
  valign = 1,
  width = 0.25
)
}
\arguments{
\item{image}{The image to place. Can be a file path, a URL, or a raw vector with image data,
as in \code{magick::image_read()}. Can also be an image previously created by \code{magick::image_read()} and
related functions.}

\item{x}{The x location of the image. (Left side if \code{hjust = 0}.)}

\item{y}{The y location of the image. (Bottom side if \code{vjust = 0}.)}

\item{hjust}{Horizontal and vertical justification relative to x.}

\item{vjust}{Horizontal and vertical justification relative to x.}

\item{halign}{Horizontal and vertical justification of the image inside
the box.}

\item{valign}{Horizontal and vertical justification of the image inside
the box.}

\item{width}{Width of the image.}
}
\description{
Places an image on the drawing canvas. This function is not intended to be
used on its own, but instead should be added to a ggplot2 plot to add a logo
in the bottom right of the plot. Because of some implementation details, this
function must be the last call in a ggplot2 chain; you won't be able to add
additional ggplot2 elements after adding the logo.
}
\details{
This function requires the \code{magick} package.
}
\examples{
library("ggplot2")

ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  logo_image()
}
