% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-sage-hue.R
\name{sage_hue_pal}
\alias{sage_hue_pal}
\title{Sage hue palette (discrete)}
\usage{
sage_hue_pal(level = "500", direction = 1)
}
\arguments{
\item{level}{Color intensity. One of "200", "300", "400", "500", "600",
"700", "800".}

\item{direction}{If -1, reverses the default order of the palette}
}
\value{
A function that gives the hex colors for the requested number of
}
\description{
Sage hue palette (discrete)
}
\examples{
sage_hue_pal()(5)
sage_hue_pal(direction = -1)(5)
sage_hue_pal(level = "800")(5)
}
