% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-colour-sage-b.R
\name{scale_colour_sage_b}
\alias{scale_colour_sage_b}
\alias{scale_fill_sage_b}
\alias{scale_color_sage_b}
\title{Binned color scales}
\usage{
scale_colour_sage_b(
  ...,
  option = "royal",
  low = NULL,
  high = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "colour"
)

scale_fill_sage_b(
  ...,
  option,
  low = NULL,
  high = NULL,
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill"
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_colour_gradient]{ggplot2::scale_colour_gradient}}
  \describe{
    \item{\code{space}}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}
  }}

\item{option}{One of the Sage colors. See \code{names(sage_colors)} for a list.}

\item{low}{Colours for low and high ends of the gradient.}

\item{high}{Colours for low and high ends of the gradient.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}
\description{
These create a binned gradient of colors based on the high and low value of a
given hue from within the list of Sage branded colors.
}
\examples{
library("ggplot2")

df_na <- data.frame(
  value = seq(1, 20),
  x = runif(20),
  y = runif(20),
  z1 = c(rep(NA, 10), rnorm(10))
)

ggplot(df_na, aes(x, y)) +
  geom_point(aes(colour = z1)) +
  scale_colour_sage_b(option = "apple")

ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_sage_b(option = "powder")
}
