% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sage.R
\name{theme_sage}
\alias{theme_sage}
\title{Sage ggplot2 theme}
\usage{
theme_sage(
  base_size = 14,
  base_family = "Lato",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A theme that uses Lato as the base font family (if available) and increases
the base font size.
}
\examples{
\dontrun{

library("ggplot2")

ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  scale_color_sage_d() +
  theme_sage()
}
}
