% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{sage_colors}
\alias{sage_colors}
\title{Complete list of Sage colors}
\format{
An object of class \code{list} of length 14.
}
\usage{
sage_colors
}
\description{
Complete list of Sage colors
}
\examples{
head(sage_colors, n = 3)
invisible(lapply(sage_colors, scales::show_col))
}
\keyword{datasets}
