% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-sage-gradient.R
\name{sage_gradient_pal}
\alias{sage_gradient_pal}
\title{Sage gradient palette}
\usage{
sage_gradient_pal(option = "royal", low = NULL, high = NULL)
}
\arguments{
\item{option}{One of the Sage colors. See \code{names(sage_colors)} for a list.}

\item{low}{Optional starting value (if \code{NULL}, will default to the \code{"200"}
value for the chosen color option)}

\item{high}{Optional ending value (if \code{NULL}, will default to the \code{"800"}
value for the chosen color option)}
}
\value{
A function that can be called on values between 0 and 1 to return the
hex color of the gradient at that point in the range.
}
\description{
Used for both continuous and binned scales, this function creates a gradient
of colors from the given high and low hex codes (or, by default, from the
"200" to "800" values for a chosen color option).
}
\examples{
sage_gradient_pal()(1)
sage_gradient_pal()(0.5)
sage_gradient_pal(
  option = "powder",
  low = sage_colors[["powder"]][["300"]],
  high = sage_colors[["powder"]][["700"]]
)(1)
}
