% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sage.R
\name{theme_sage}
\alias{theme_sage}
\title{Sage ggplot2 theme}
\usage{
theme_sage(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:theme_grey]{ggplot2::theme_grey()}}}
}
\description{
Right now just replaces the base font family with Lato and increases the base
font size.
}
\examples{
\dontrun{

library("ggplot2")

# Requires Lato font to be installed
ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  scale_color_sage_d() +
  theme_sage()
}
}
